blinn-phong: fix specular highlights
Previously if you translated a shape -- or made any other change to its input
coord -- the specular highlights would be the same as the untranslated shape.
Now we compute specular based on the actual final position in the world -- the
same position that we were already using to compute light intensities -- which
is obviously the correct thing to do.