unthinkable performance improvements
We no longer recompile the entire shader from scratch whenever we move
the camera. Now the camera position and rotation matrix is a uniform,
and we just set it and re-render whenever you change any view properties.

This is such a massive improvement when using noise functions -- compiling
a shader with the perlin3 implementation takes over 15ms on my laptop.

Probably also nice to precompute the camera matrix, but that's so tiny compared
to the other perf win that you'd never even notice it.