fix the way rotation works with multiple axis pairs
So to rotate a shape by some amount, you need to rotate space by the
opposite amount. And that wasn't what we were doing: we were rotating space
by negative angles, which works when you only have one axis of rotation, but
if you have multiple axes of rotation you also need to reverse the order that
you do each rotation. Which we weren't.

So instead of negating the angles and reversing the order, we can just
right-multiply by the rotation matrix instead, and it now behaves exactly like
we'd expect it to.