fix build script on macos
macOS Tahoe ships a new and apparently much more strict version of sed that
caused some of the sed invocations in the build scripts to fail. Update them
to appease the new verison, and then also switch to GNU sed in the nix shell
so that I never have to think about that again.